Fix configuration of multiple launch modes
The 'pydantic.BaseModel' used as value in the 'configs' dict of
the 'ProductConfig' caused the configuration to be discarded when
loading from file (cast to 'BaseModel').
This was not captured by the existing tests because they were not
resetting the configuration between calls to the 'configure' CLI
command.

This is fixed by accepting 'Any', and only casting to the specific
type when the configuration is actually requested.

An alternative would be dynamically constructing the pydantic model
for the configuration from plugins. Besides being overly complicated,
this has the drawback that the configuration is lost if a plugin
is uninstalled.