Replace pydantic by dataclasses (#33)
The `pydantic` models have a rather large API, which users will inevitably come
to rely on. To avoid this, we switch to using `dataclass` for the configuration.

The main drawback of using `dataclass` is that there's no standardized way to specify the 
description, which is now passed as a `metadata` value.
Another notable difference is that `dataclass` doesn't do run-time type checking.