Remove use of dataclasses.MISSING in docs
When creating dataclasses, non-default arguments must come after
default arguments. This means the dynamic use of dataclasses.MISSING
is problematic; there may not be an order of arguments that ensures
this property.
For this reason, we remove use of dataclasses.MISSING from the
docs (so as to not serve as a bad example).