Add an example plugin
Add a complete example plugin package ``example_httpserver_plugin``,
which is used to demonstrate the common configuration and launch
operations in the examples section.

Use sphinx-gallery to render the examples.

Fix the case where the type annotations in the launcher config are
defined with delayed annotations (``from __future__ import annotations``)
and are therefore represented as strings instead of the actual type.